Commented out base declaration
The base is added to vite.config.js so that gh-pages deploys the build,
the index.html will have the proper URL for GH Pages. But during dev,
the "/DeleGate/" added to the end of the localhost causes the routing to
break. This is because no routes are setup for "/DeleGate" nor do we
want to.

The solution: New deployment pipeline. I will be working on this over
the next few weeks.

The temporary fix: in `vite.config.js` comment out `base: "/DeleGate/"
and continue development as normal. When ready to deploy, add the line
back in and run the deploy command.