Added redirect to frontend on successful login
In the current implementation, Spring Security wants to redirect to a
local route (relative to the server at 8080) instead of where our
frontend is being hosted (at 5173). I simply added a GetMapping to the
route the success controller was sending us too, which then directs the
user to http://localhost:5173.