Implemented routing to server for login
For dev, we need to route to the server at port 8080, as this
isn't handled for us via the host. I setup Vite's proxy option
in the config to create routes that lead to the server.

Updated the login buttons to match the new routes.

TODO: Add route protection and detecting successful logins.